update with more documentation
